backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
Deprecated==1.2.18
distro==1.9.0
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Limiter==3.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
limits==5.1.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
openai==1.79.0
ordered-set==4.1.0
packaging==25.0
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
python-dotenv==1.0.0
requests==2.31.0
rich==13.9.4
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
gunicorn==21.2.0
gevent==24.2.1
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
psycopg2-binary==2.9.9
flask-jwt-extended==4.6.0
flask-bcrypt==1.0.1